test crc.c x86_64 compile fixes
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@413 048eaad1-c872-4173-8c30-523bb29f53ec

+++ crc.h
@@ -25,7 +25,7 @@
 
 #if defined(CRC_CCITT)
 
-typedef uint32_t  crc;
+typedef uint16_t  crc;
 
 #define CRC_NAME			"CRC-CCITT"
 #define POLYNOMIAL			0x1021
